Move incorrect case diagnostic things into their module
